Configuration via `typesafe` and `ficus`
All the original hardcoded values in the `Docker*Service` traits are
now in the `application.conf` in `core/test/resources`.

The new Typesafe config class has case classses to read the config and
transform that data into a `DockerContainer`.

These case classes have the defaults that were in the
`DockerReadyChecker`.

`DockerKit` loads the config and has a method that takes a path to
load.  `scalatest` and `specs2` tests run well.

One change in the config is that the `Kafka` service needs a `DOCKER_IP`
environmental variable that can be resolved in the typesafe config.